requirements.txt
langchain
openai
pinecone-client
tiktoken
PyPDF2
docx2pdf
pypdf
docx2txt